fail commit
program is not what its supposed to be. its supposed to determine whether the number generated by the user is in the list. but what it really does is see if the numbers are in the if statement then return. basically im supposed to be making use of the list but im not.